big changes to controllers

+++ backend/src/main/java/com/pcc/pc_configurator/repositories/CaseRepository.java
@@ -4,5 +4,5 @@ import com.pcc.pc_configurator.entities.Case_;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CaseRepository extends JpaRepository<Case_,Long> {
+public interface CaseRepository extends JpaRepository<Case_,Integer> {
 }
